Clean skin retrieval and media link
Use the DokuHTTPClient class to fetch the skin to benefit from
dokuwiki's proxy settings and user preferences.
End the media link with ".png" so that the different functions and the
browser better recognise that the link points to an image. As the link
already contains the '.png', the _getSkin function do not add the
extension anymore.